vision-engine/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
requests==2.32.5
Pillow==11.3.0
numpy==2.3.2
psycopg[binary]==3.2.10
pgvector==0.4.1
open_clip_torch==2.32.0
torch==2.8.0

